+++ Sources/SuperwallKit/Config/ConfigManager.swift
@@ -4,6 +4,7 @@
 //
 //  Created by Yusuf Tör on 22/06/2022.
 //
+// swiftlint:disable type_body_length
 
 import UIKit
 import Combine
@@ -26,7 +27,7 @@ class ConfigManager {
   }
 
   /// Options for configuring the SDK.
-  var options = SuperwallOptions()
+  var options: SuperwallOptions
 
   /// A dictionary of triggers by their event name.
   @DispatchQueueBacked
@@ -42,6 +43,7 @@ class ConfigManager {
   private unowned let storage: Storage
   private unowned let network: Network
   private unowned let paywallManager: PaywallManager
+  private unowned let deviceHelper: DeviceHelper
 
   /// A task that is non-`nil` when preloading all paywalls.
   private var currentPreloadingTask: Task<Void, Never>?
@@ -49,20 +51,20 @@ class ConfigManager {
   private let factory: RequestFactory & RuleAttributesFactory & ReceiptFactory
 
   init(
-    options: SuperwallOptions?,
+    options: SuperwallOptions,
     storeKitManager: StoreKitManager,
     storage: Storage,
     network: Network,
     paywallManager: PaywallManager,
+    deviceHelper: DeviceHelper,
     factory: RequestFactory & RuleAttributesFactory & ReceiptFactory
   ) {
-    if let options = options {
-      self.options = options
-    }
+    self.options = options
     self.storeKitManager = storeKitManager
     self.storage = storage
     self.network = network
     self.paywallManager = paywallManager
+    self.deviceHelper = deviceHelper
     self.factory = factory
   }
 
@@ -70,9 +72,12 @@ class ConfigManager {
     do {
       _ = await factory.loadPurchasedProducts()
 
-      let config = try await network.getConfig { [weak self] in
+      async let configRequest = network.getConfig { [weak self] in
         self?.configState.send(.retrying)
       }
+      async let geoRequest: Void = deviceHelper.getGeoInfo()
+
+      let (config, _) = try await (configRequest, geoRequest)
 
       Task { await sendProductsBack(from: config) }
 
@@ -287,8 +292,16 @@ class ConfigManager {
             presentationSourceType: nil,
             retryCount: 6
           )
-          _ = try? await self.paywallManager.getPaywallViewController(
-            from: request,
+
+          guard let paywall = try? await paywallManager.getPaywall(from: request) else {
+            return
+          }
+
+          await self.paywallManager.attemptToPreloadArchive(from: paywall)
+
+          _ = try? await self.paywallManager.getViewController(
+            for: paywall,
+            isDebuggerLaunched: request.isDebuggerLaunched,
             isForPresentation: true,
             isPreloading: true,
             delegate: nil